Correccion de TarjetaCredito , se inserta bien en la base de datos

+++ src/main/java/org/grupo4/practica_integradora_g4/repositories/TarjetaCreditoRepository.java
@@ -3,5 +3,5 @@ package org.grupo4.practica_integradora_g4.repositories;
 import org.grupo4.practica_integradora_g4.model.entidades.TarjetaCredito;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TarjetaCreditoRepository extends JpaRepository<TarjetaCredito, Long> {
+public interface TarjetaCreditoRepository extends JpaRepository<TarjetaCredito, Integer> {
 }
